*Misc updates.

+++ src/map/buyingstore.c
@@ -630,7 +630,7 @@ void do_init_buyingstore_autotrade(void) {
 				// Initialize player
 				CREATE(at->sd, struct map_session_data, 1);
 				pc_setnewpc(at->sd, at->account_id, at->char_id, 0, gettick(), at->sex, 0);
-				at->sd->state.autotrade = 2|8;
+				at->sd->state.autotrade = (unsigned int)(2|8);
 				at->sd->state.monster_ignore = (battle_config.autotrade_monsterignore);
 				chrif_authreq(at->sd, true);
 				uidb_put(buyingstore_autotrader_db, at->char_id, at);
